Remove IdbStore and update tests
