+++ Cargo.toml
@@ -60,13 +60,6 @@ features = [
     "Crypto",
     "DomException",
     "Headers",
-    "IdbDatabase",
-    "IdbFactory",
-    "IdbObjectStore",
-    "IdbOpenDbRequest",
-    "IdbTransaction",
-    "IdbTransactionMode",
-    "IdbVersionChangeEvent",
     "Request",
     "RequestInit",
     "RequestMode",
